Update README.md to add alternate subscribe link
Added alternate subscribe link for situations where the user needs ot use something other than the built in RSS link provided by Ghost. The subecription button is changed to the alternate subscribe link during page load. It does not change the server-side output for the alternate link for the site.